+++ cpp/src/utils/CommonUtil.h
@@ -11,7 +11,7 @@
 #include "Error.h"
 
 namespace zilliz {
-namespace vecwise {
+namespace milvus {
 namespace server {
 
 class CommonUtil {