+++ cpp/src/utils/StringHelpFunctions.cpp
@@ -6,7 +6,7 @@
 #include "StringHelpFunctions.h"
 
 namespace zilliz {
-namespace vecwise {
+namespace milvus {
 namespace server {
 
 void StringHelpFunctions::TrimStringLineBreak(std::string &string) {