+++ cpp/CHANGELOG.md
@@ -7,6 +7,8 @@ Please mark all change in change log and use the ticket from JIRA.
 ## Bug
 
 ## Improvement
+- MS-82 - Update server startup welcome message
+- MS-83 - Update vecwise to Milvus
 
 ## New Feature